week-8/app/requirements.txt
Authlib==1.6.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.2
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
